socket: setup event handlers for server and client
Create an index.js file to hold the javascript code from the client side.
Event handlers such as createMessage emitted from the client side and listened
on the server side, as well as newMessage emitted from the server and received
on the client side.

Also refactoring arrow functions in the client for compatibility with browsers.

Ref.: Lectures 108 and 109

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>