chat-app: add geolocation support via geolocation api
This commit inplements a geolocation mechanism using the mozilla's geolocation
api to send a link with a map indicating the user's position, after they press
a 'Send Location' button.
Note that the implementation requires the connection protocol to be secure via https.

Also adding tests for this functionality.

Ref. Lectures 114 and 115.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>